[HEIDELPAY-212] Adjust register webhook command - Add logger for WebhookHandler - Adjust `TransitionEventListener` - Rebuild js files

+++ src/Commands/RegisterWebhookCommand.php
@@ -6,12 +6,18 @@ namespace HeidelPayment6\Commands;
 
 use HeidelPayment6\Components\ClientFactory\ClientFactoryInterface;
 use heidelpayPHP\Exceptions\HeidelpayApiException;
+use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\System\SalesChannel\Aggregate\SalesChannelDomain\SalesChannelDomainEntity;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+use Symfony\Component\Routing\RequestContext;
 use Symfony\Component\Routing\Router;
 use Throwable;
 
@@ -28,10 +34,17 @@ class RegisterWebhookCommand extends Command
     /** @var Router */
     private $router;
 
-    public function __construct(ClientFactoryInterface $clientFactory, Router $router)
-    {
-        $this->clientFactory = $clientFactory;
-        $this->router        = $router;
+    /** @var EntityRepositoryInterface */
+    private $domainRepository;
+
+    public function __construct(
+        ClientFactoryInterface $clientFactory,
+        Router $router,
+        EntityRepositoryInterface $domainRepository
+    ) {
+        $this->clientFactory    = $clientFactory;
+        $this->router           = $router;
+        $this->domainRepository = $domainRepository;
 
         parent::__construct();
     }
@@ -51,23 +64,38 @@ class RegisterWebhookCommand extends Command
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $style = new SymfonyStyle($input, $output);
+        $style        = new SymfonyStyle($input, $output);
+        $providedHost = $input->getArgument('host');
+        $parsedHost   = parse_url($input->getArgument('host'));
+
+        if (empty($parsedHost['host']) || empty($parsedHost['scheme'])) {
+            $style->warning('The provided host is invalid.');
+
+            return self::EXIT_CODE_INVALID_HOST;
+        }
+
+        if (!$this->domainExists($providedHost)) {
+            $style->warning('The provided host does not exist in any saleschannel.');
+
+            $possibleDomains = [];
+            /** @var SalesChannelDomainEntity $salesChannelDomain */
+            foreach ($this->domainRepository->search(new Criteria(), Context::createDefaultContext()) as $salesChannelDomain) {
+                $possibleDomains[] = [$salesChannelDomain->getUrl()];
+            }
+            $style->table(['Possible domains'], $possibleDomains);
+
+            return self::EXIT_CODE_INVALID_HOST;
+        }
 
         try {
-            $client = $this->clientFactory->createClient();
-            $client->deleteAllWebhooks();
+            $context = $this->getContext($parsedHost);
 
-            $host = parse_url($input->getArgument('host'));
-
-            if (empty($host['host']) || empty($host['scheme'])) {
-                $style->warning('The provided host is invalid.');
-
-                return self::EXIT_CODE_INVALID_HOST;
+            if (!$context) {
+                return self::EXIT_CODE_UNKNOWN_ERROR;
             }
 
-            $context = $this->router->getContext();
-            $context->setHost($host['host']);
-            $context->setScheme($host['scheme']);
+            $client = $this->clientFactory->createClient();
+            $client->deleteAllWebhooks();
 
             $url = $this->router->generate('heidelpay.webhook.execute', [], UrlGeneratorInterface::ABSOLUTE_URL);
 
@@ -86,5 +114,35 @@ class RegisterWebhookCommand extends Command
         }
 
         return self::EXIT_CODE_SUCCESS;
+    }
+
+    protected function domainExists(string $url): bool
+    {
+        $domainCriteria = new Criteria();
+        $domainCriteria->addFilter(new EqualsFilter('url', $url));
+
+        $salesChannelResult = $this->domainRepository->search($domainCriteria, Context::createDefaultContext());
+        /** @var null|SalesChannelDomainEntity $firstResult */
+        $firstResult = $salesChannelResult->first();
+
+        if (empty($firstResult)) {
+            return false;
+        }
+
+        return true;
+    }
+
+    protected function getContext($host): ?RequestContext
+    {
+        $context = $this->router->getContext();
+
+        if (!$context) {
+            return null;
+        }
+
+        $context->setHost($host['host']);
+        $context->setScheme($host['scheme']);
+
+        return $context;
     }
 }
